Clarify service-layer command naming and move driver config to config.
Unify HTTP/WS/remote JSON on CommandRequest with dispatch_command routing,
ParsedCommandRequest/CommandError types, and driver options beside AppConfig.
