+++ src/service/mod.rs
@@ -1,34 +1,37 @@
-//! Application service: wire-format validation and UI-style command dispatch.
+//! Application service: command validation and UI-style dispatch.
 
 use anyhow::{Context, Result};
+use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 use std::sync::Arc;
 
+use crate::config::DriverKind;
 use crate::controller::BlindController;
 use crate::core::{Channel, Command};
-use crate::driver::{CommandOutcome, DriverKind, SelectedChannelRx, TELIS_PROG_UNAVAILABLE};
+use crate::driver::{CommandOutcome, SelectedChannelRx, TELIS_PROG_UNAVAILABLE};
 
-/// Parsed HTTP/WebSocket/CLI press ready for dispatch.
+/// Parsed command ready for dispatch.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PressRequest {
+pub struct ParsedCommandRequest {
     pub command: Command,
     pub channel: Option<Channel>,
 }
 
-/// Wire-format command body (HTTP JSON / WebSocket text).
-#[derive(Debug, Clone)]
-pub struct WirePress {
+/// HTTP/JSON command body (`POST /command`, WebSocket text, CLI remote POST).
+#[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
+pub struct CommandRequest {
     pub command: String,
     pub channel: Option<Channel>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PressError {
+pub enum CommandError {
     Invalid(String),
     PairingUnavailable,
 }
 
-impl std::fmt::Display for PressError {
+impl std::fmt::Display for CommandError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::Invalid(msg) => write!(f, "{msg}"),
@@ -37,10 +40,10 @@ impl std::fmt::Display for PressError {
     }
 }
 
-impl std::error::Error for PressError {}
+impl std::error::Error for CommandError {}
 
-fn press_error(err: anyhow::Error) -> PressError {
-    PressError::Invalid(format!("{err:?}"))
+fn command_error(err: anyhow::Error) -> CommandError {
+    CommandError::Invalid(format!("{err:?}"))
 }
 
 /// Central dispatch for REST, WebSocket, and in-process callers.
@@ -67,42 +70,54 @@ impl BlindService {
         self.controller.subscribe_selection()
     }
 
-    /// Validate a wire-format press. Does not touch hardware.
-    pub fn parse_wire(wire: WirePress) -> Result<PressRequest, PressError> {
-        let WirePress { command, channel } = wire;
-        let cmd = Command::from_str(&command).map_err(|e| PressError::Invalid(e.to_string()))?;
+    /// Validate a command request. Does not touch hardware.
+    pub fn parse_command(request: CommandRequest) -> Result<ParsedCommandRequest, CommandError> {
+        let CommandRequest { command, channel } = request;
+        let cmd = Command::from_str(&command).map_err(|e| CommandError::Invalid(e.to_string()))?;
         let channel = match (cmd, channel) {
             (Command::Prog | Command::ProgLong, Some(channel)) => Some(channel),
             (Command::Prog | Command::ProgLong, None) => {
-                return Err(PressError::Invalid(
+                return Err(CommandError::Invalid(
                     "prog and prog_long require a channel".to_string(),
                 ));
             }
             (Command::Select, channel) => channel,
             (Command::Up | Command::Down | Command::Stop, channel) => channel,
         };
-        Ok(PressRequest {
+        Ok(ParsedCommandRequest {
             command: cmd,
             channel,
         })
     }
 
     /// Reject pairing commands when the active driver cannot transmit them.
-    pub fn ensure_pairing_allowed(&self, command: Command) -> Result<(), PressError> {
+    pub fn ensure_pairing_allowed(&self, command: Command) -> Result<(), CommandError> {
         Self::ensure_pairing_for_kind(self.kind, command)
     }
 
-    pub fn ensure_pairing_for_kind(kind: DriverKind, command: Command) -> Result<(), PressError> {
+    pub fn ensure_pairing_for_kind(kind: DriverKind, command: Command) -> Result<(), CommandError> {
         if matches!(command, Command::Prog | Command::ProgLong) && !kind.supports_pairing() {
-            return Err(PressError::PairingUnavailable);
+            return Err(CommandError::PairingUnavailable);
         }
         Ok(())
     }
 
-    /// UI-style dispatch: optional channel selects before directional commands.
-    pub async fn press(&self, request: PressRequest) -> Result<CommandOutcome, PressError> {
+    /// Validate and dispatch a command. `select` runs directly; directional commands
+    /// optionally select a channel first, then route to the controller.
+    pub async fn dispatch_command(
+        &self,
+        request: CommandRequest,
+    ) -> Result<CommandOutcome, CommandError> {
+        self.dispatch_parsed_command(Self::parse_command(request)?)
+            .await
+    }
+
+    async fn dispatch_parsed_command(
+        &self,
+        request: ParsedCommandRequest,
+    ) -> Result<CommandOutcome, CommandError> {
         self.ensure_pairing_allowed(request.command)?;
-        let PressRequest {
+        let ParsedCommandRequest {
             command: cmd,
             channel,
         } = request;
@@ -112,7 +127,7 @@ impl BlindService {
                 .execute(cmd, channel)
                 .await
                 .context("executing select command")
-                .map_err(press_error);
+                .map_err(command_error);
         }
 
         if let Some(channel) = channel {
@@ -120,29 +135,28 @@ impl BlindService {
                 .execute(Command::Select, Some(channel))
                 .await
                 .context("selecting channel before command")
-                .map_err(press_error)?;
+                .map_err(command_error)?;
         }
         self.controller
             .execute(cmd, None)
             .await
             .with_context(|| format!("executing {cmd:?} command"))
-            .map_err(press_error)
-    }
-
-    pub async fn press_wire(&self, wire: WirePress) -> Result<CommandOutcome, PressError> {
-        let request = Self::parse_wire(wire)?;
-        self.press(request).await
+            .map_err(command_error)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::config::DriverKind;
     use crate::controller::BlindController;
-    use crate::driver::{DriverKind, ProtocolOperation};
+    use crate::driver::ProtocolOperation;
 
-    fn parse(command: &str, channel: Option<Channel>) -> Result<PressRequest, PressError> {
-        BlindService::parse_wire(WirePress {
+    fn parse(
+        command: &str,
+        channel: Option<Channel>,
+    ) -> Result<ParsedCommandRequest, CommandError> {
+        BlindService::parse_command(CommandRequest {
             command: command.to_string(),
             channel,
         })
@@ -153,7 +167,7 @@ mod tests {
         assert!(!DriverKind::Telis.supports_pairing());
         assert!(matches!(
             BlindService::ensure_pairing_for_kind(DriverKind::Telis, Command::Prog),
-            Err(PressError::PairingUnavailable)
+            Err(CommandError::PairingUnavailable)
         ));
     }
 
@@ -187,7 +201,7 @@ mod tests {
     #[test]
     fn parse_rejects_prog_without_channel() {
         let err = parse("prog", None).unwrap_err();
-        assert!(matches!(err, PressError::Invalid(_)));
+        assert!(matches!(err, CommandError::Invalid(_)));
         assert!(err.to_string().contains("require a channel"));
     }
 
@@ -206,17 +220,17 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn press_with_channel_selects_then_executes() {
+    async fn dispatch_command_with_channel_selects_then_executes() {
         let controller = Arc::new(
-            BlindController::with_driver(crate::driver::DriverConfig::fake())
+            BlindController::with_driver(crate::config::DriverConfig::fake())
                 .await
                 .unwrap(),
         );
         let blinds = Arc::new(BlindService::new(controller.clone(), DriverKind::Fake));
 
         blinds
-            .press(PressRequest {
-                command: Command::Up,
+            .dispatch_command(CommandRequest {
+                command: "up".to_string(),
                 channel: Some(Channel::L3),
             })
             .await
@@ -233,5 +247,29 @@ mod tests {
                 },
             ]
         );
+    }
+
+    #[test]
+    fn command_request_accepts_channel_field() {
+        let req: CommandRequest =
+            serde_json::from_str(r#"{"command":"up","channel":"L1"}"#).unwrap();
+
+        assert_eq!(req.command, "up");
+        assert_eq!(req.channel, Some(Channel::L1));
+    }
+
+    #[test]
+    fn command_request_accepts_prog_long() {
+        let req: CommandRequest =
+            serde_json::from_str(r#"{"command":"prog_long","channel":"L1"}"#).unwrap();
+        assert_eq!(req.command, "prog_long");
+        assert_eq!(req.channel, Some(Channel::L1));
+    }
+
+    #[test]
+    fn command_request_rejects_legacy_led_field() {
+        let err = serde_json::from_str::<CommandRequest>(r#"{"command":"select","led":"L1"}"#)
+            .unwrap_err();
+        assert!(err.to_string().contains("unknown field"));
     }
 }
